Move toggle logic to Searcher component, delete utils directory
